CSS and layout fixed.
