+++ app/layouts/Content.tsx
@@ -6,12 +6,15 @@ import Footer from '@/layouts/Footer'
 
 const useStyles = makeStyles((theme: Theme) => ({
   content: {
+    maxWidth: 'none',
     flexGrow: 1,
     height: '100vh',
     overflow: 'auto',
-
     [theme.breakpoints.down('md')]: {
-      paddingTop: theme.spacing(6),
+      paddingTop: theme.spacing(12),
+    },
+    [theme.breakpoints.up('md')]: {
+      paddingTop: theme.spacing(3),
     },
   },
 }))
@@ -21,7 +24,6 @@ interface IContent {
 
 const Content = ({ children }: IContent): React.ReactElement => {
   const classes = useStyles()
-
   return (
     <Container component="main" maxWidth="lg" fixed className={classes.content}>
       {children}
